+++ example/src/test/java/com/alipay/sofa/rpc/grpc/GreeterImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.grpc.registry;
+package com.alipay.sofa.rpc.grpc;
 
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
